requirements.txt
PyQt6==6.4.0
onnxruntime==1.12.1
opencv-python==4.5.5.64
folium==0.12.1
Pillow==9.3.0
numpy==1.21.5
openpyxl==3.0.10